requirements.txt
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
Django==3.1.2
django-nose==1.4.7
django-utils-six==2.0
djangorestframework==3.12.1
idna==2.10
nose==1.3.7
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
